job_finder/requirements.txt
django
django-filter
djangorestframework
psycopg2-binary
httpx
pytest
pytest-django